lsp-devtools: refactor modules
