+++ lib/lsp-devtools/lsp_devtools/agent/__init__.py
@@ -13,16 +13,13 @@ from .agent import MessageHeader
 from .agent import MessageSource
 from .client import AgentClient
 from .server import AgentServer
-from .server import JsonRPCHandler
-from .server import JsonRPCMessage
 
 __all__ = [
     "Agent",
     "AgentClient",
     "AgentServer",
+    "MessageHeader",
     "MessageSource",
-    "JsonRPCHandler",
-    "JsonRPCMessage",
 ]
