+++ lib/lsp-devtools/lsp_devtools/handlers/sql.py
@@ -1,19 +1,22 @@
+from __future__ import annotations
+
 import json
 import pathlib
 import sqlite3
+import typing
 from contextlib import closing
+from datetime import datetime
 from importlib import resources
 
-from lsp_devtools.handlers import LspHandler
-from lsp_devtools.handlers import LspMessage
+from .jsonrpc import JsonRPCHandler
+from .jsonrpc import JsonRPCMessage
 
 
-class SqlHandler(LspHandler):
-    """A logging handler that sends log records to a SQL database"""
+class SqlHandler(JsonRPCHandler):
+    """A handler that sends messages to a SQL database"""
 
     def __init__(self, dbpath: pathlib.Path, *args, **kwargs):
         super().__init__(*args, **kwargs)
-
         self.dbpath = dbpath
 
         resource = resources.files("lsp_devtools.handlers").joinpath("dbinit.sql")
@@ -22,21 +25,25 @@ class SqlHandler(LspHandler):
         with closing(sqlite3.connect(self.dbpath)) as conn:
             conn.executescript(sql_script)
 
-    def handle_message(self, message: LspMessage):
+    def handle(self, message: JsonRPCMessage):
         with closing(sqlite3.connect(self.dbpath)) as conn:
             cursor = conn.cursor()
             cursor.execute(
-                "INSERT INTO protocol VALUES (?, ?, ?, ?, ?, ?, ?, ?)",
+                "INSERT INTO messages VALUES (?, ?, ?)",
                 (
-                    message.session,
-                    message.timestamp,
-                    message.source,
-                    message.id,
-                    message.method,
-                    json.dumps(message.params) if message.params else None,
-                    json.dumps(message.result) if message.result else None,
-                    json.dumps(message.error) if message.error else None,
+                    json.dumps(message.metadata, default=to_json),
+                    json.dumps(message.headers, default=to_json),
+                    json.dumps(message.body, default=to_json),
                 ),
             )
 
             conn.commit()
+
+
+def to_json(o):
+    """Convert unserializable types to a JSON compatible representation"""
+
+    if isinstance(o, datetime):
+        return o.isoformat(" ")
+
+    raise ValueError(f"Unknown type {o.__class__.__name__!r}")
